requirements.txt
Flask==2.2.3
Werkzeug==2.3.8
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
PyMySQL==1.0.2
requests==2.28.2
qrcode[pil]==7.4.2
Pillow==9.5.0
